The very start of the repository
Previously made a start on splitting the prettylst script
into modules to attempt to make it more understandable and
maintainable.

The Options Package and the Logging package are from that.

Everything else I'm starting again in the hope of keeping better
control over which module is reponsible for which function.