Change processing of data defined tags
Some tags such as FACT are totally data defined, it is not useful
to add the subtags explicitly to the validation.

Others are standard and need their own header and sorting order.
FACT:BaseSize for example.

When an non-standard subtag is identified, an info message is printed
this is mostly invisible since it is below the default logging level