General tidying
Renamed some variables.

Moved some code to better organise what happens when, the script now
checks for all the options that might make it exit eraly before it does
anything else.

Deleted some data variables that are now in other places.

Removed some if statements at the end of my declarations, the perl
documentation explicitly says this is undefined behaviour and should not
be used.
a