Tidy up
Move more of the reporting into LstTidy::Report. Rewrite some almost
identical reporting functions to just be report.

Move the PREXXX and BONUS report into LstTidy Report.

Changed a lot of the file collecting hashes to be the %file hash,
eliminated some useless if statments. The script now makes a categorised
list of all the files it might be interested in. These are then availble
if needed.

Reindented more of the script.