Major changes.
Tag has become Token and most of what was in Validate is now part
of the object.

All the validity data and tag ordering has moved to a new Data.pm
this cuts down on modules using each other, they all use data instead.

Fixed the DATACONTROL lines. There are now three types of line
in DATACONTROL with suitable header info, ordering, and validation.

Fixed the comma in entity name warning, now warns for commas in key
or in entity name if there is no key.