Tidy up tidylst.pl
Tidy up comments and correct spelling.  Redo the perldoc. Tidy up html
generation. Tidy up the generated css.

Remove unused imports from Data.

Import makeExportListString and printToExportList from Report. Use them
for PCC files.

Rename filelist and filename to pccList and pccName respectively.

Check for GAME to GAMEMODE conversion before checking GAMEMODE.

Implement the logic to make sure there is a single reformatted by header
at the top of the PCC.

Elsewhere
---------
Remove the xcheck option, it's on by default which makes it redundant).

Remove the vestiges of the oldsourcetag option, tidylst is for modern
lst and doesn't need it.

Fix the bug in source line replacement and turn it back on.

Move makeExportListString to Report and export it.

Export printToExportList from Report and use it in Parse