Move Things out of tidylst.pl
Extract teh file processing logic from the loop and make it an operatin
in Parse.pm.  This also means we had to move createDir to Parse.

We need the File Header string and regexs in both Parse and the script,
so move them to Data to amke the universally accessible.