Expand the validation
Clean up the adding of abilities to the validiation data. Have
eliminated all the spurious this ability doesn't exist errors.

Add game mode to the file type data, amy use this later when parsing the
validstats etc. to ensure that each gamemode is independent.

Move all of the add valid entity stuff to Parse.pm