Fix reformatting bug
SUB lines in biosettings were not getting their own formatter.
This caused them to be blank.

Refactored the calculation of required padding out to a seaparete
operation which is called in multiple places.

Added logic to ensure that all of the lines that are First column
formatted have the smae first column length throughout a file (looks
neater).