Change the handling of filenames
Add in a File object to hold the name as it existed in the PCC file, the
actual file we will read, a vendor path if that is needed, and the
output name.

A check has been added to ensure that only files actually under the
input path will be written. If the PCC references files from another
directory tree, these will be check but not rewritten.