Fix some bugs
Vendor path wasn't getting processed correctly. Now correctly identifies
a file in the vendorpath and doesn't look for it under the inputpath.

Correct the spelling of SHIELDPROF so that its header lines populate
correctly.

Correct the block detection so that blocks of lines that shuld be
formatted together are actually formatted together.